menuのeditとnewをコンポネートで冗長
