+++ src/router/index.js
@@ -7,7 +7,9 @@ import Genres from '@/components/Genres'
 
 import Menus from '@/components/menus/index'
 import MenuNew from '@/components/menus/new'
+import MenuEdit from '@/components/menus/edit'
 import MenuShow from '@/components/menus/show'
+// import MenuForm from '@/components/menus/_form'
 
 import Users from '@/components/users/index.vue'
 import UserNew from '@/components/users/new.vue'
@@ -42,6 +44,7 @@ export default new Router({
       component: Genres
     },
     {
+      // ジャンル指定用のID
       path: '/menus/:id',
       name: 'Menus',
       component: Menus
@@ -50,6 +53,11 @@ export default new Router({
       path: '/menu/new',
       name: 'MenuNew',
       component: MenuNew
+    },
+    {
+      path: '/menu/edit/:id',
+      name: 'MenuEdit',
+      component: MenuEdit
     },
     {
       path: '/menu/:id',
